Typo and grammatical errors
In the first paragraph, below the first picture diagram, the line is written as, " Both view are external layout to allow a maximum of customization." This line has some errors. 
The correct line could be, " Both views **have an** external layout to allow maximum customization. 

In the usage section, in the ExpandableLayout subheading, the line is written as, "You also need to other layouts for header and content:"
The corrected line could be written as, " You also need other layouts for header and content:"

Thank you.